% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{intraplate_norm}
\alias{intraplate_norm}
\title{Intra-plate normalization}
\usage{
intraplate_norm(ds, v, ref, ext)
}
\arguments{
\item{ds}{data.table}

\item{v}{String of name of variable for values}

\item{ref}{String of name of variable for references}

\item{ext}{String giving extension given to scale factor variable name}
}
\description{
Intra-plate normalization
}
