% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization_pipeline}
\alias{normalization_pipeline}
\title{Normalize data using pipeline}
\usage{
normalization_pipeline(
  ds,
  value_name,
  steps,
  step_names,
  refs,
  extra_params = list()
)
}
\arguments{
\item{ds}{Dataset as data.table}

\item{value_name}{String of variable name containing value to be normalized}

\item{steps}{List of functions that will be used to calculate scale factors}

\item{step_names}{Names of each step provided by steps parameter}

\item{refs}{List containing a string or function that provides references to be used by each step}

\item{extra_params}{Optional parameter to provide extra parameters to functions given by steps parameter}
}
\description{
Normalize data using pipeline
}
\examples{

normalization_pipeline(
  ana.raw,
  value,
  steps = list(hyb_norm, intraplate_norm, plate_norm, calibration_norm),
  step_names = c('H', 'HI', 'HIP', 'HIPC'),
  refs = list('ref.all', derive_intraplate_ref, 'ref.calibrator', 'ref.calibrator')
)

}
